# api-docs.json
{"openapi":"3.0.1","info":{"title":"Demo Api","description":"Spring Boot RestFul API Demo by Keshav Jrall","termsOfService":"https://www.apache.org/licenses/LICENSE-2.0.txt","contact":{"name":"Keshav Jrall","url":"to be updated","email":"keshavjrall@outlook.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"Version 1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Album Controller","description":"Controller for Album and Photo Management"},{"name":"Auth Controller","description":"Controller for Account Management"}],"paths":{"/api/v1/auth/users/{user_id}/update-authorities":{"put":{"tags":["Auth Controller"],"summary":"Update authorities","operationId":"update_authorities","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthoritiesDTO"}}},"required":true},"responses":{"200":{"description":"Update Authorities","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"403":{"description":"Token/Scope Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"401":{"description":"Please check Access Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"400":{"description":"Invalid User","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/auth/profile/update-password":{"put":{"tags":["Auth Controller"],"summary":"Update Password","operationId":"update_password","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordDTO"}}},"required":true},"responses":{"403":{"description":"Token/Scope Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"401":{"description":"Please check Access Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"200":{"description":"Update Password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/albums/{album_id}/update":{"put":{"tags":["Album Controller"],"summary":"UPdate an Album","operationId":"update_album","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumPayloadDTO"}}},"required":true},"responses":{"200":{"description":"Album added","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"400":{"description":"Please enter a valid name and description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/update":{"put":{"tags":["Album Controller"],"summary":"Update photo","operationId":"update_photo","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoPayloadDTO"}}},"required":true},"responses":{"200":{"description":"Album added","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"400":{"description":"Please enter a valid name and description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/auth/users/add":{"post":{"tags":["Auth Controller"],"summary":"Add a new User","operationId":"addUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDTO"}}},"required":true},"responses":{"200":{"description":"Account Added","content":{"application/json":{"schema":{"type":"string"}}}},"401":{"description":"Please enter a valid email and Password length btw 6 and 20 characters","content":{"application/json":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/token":{"post":{"tags":["Auth Controller"],"operationId":"token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/api/v1/albums/{album_id}/upload-photos":{"post":{"tags":["Album Controller"],"summary":"Upload Photo in Album","operationId":"photos","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["files"],"type":"object","properties":{"files":{"type":"array","items":{"type":"string","format":"binary"}}}}}}},"responses":{"400":{"description":"Please check payload or token","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{"type":"array","items":{"type":"object"}}}}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/albums/add":{"post":{"tags":["Album Controller"],"summary":"Add an Album","operationId":"addAlbum","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumPayloadDTO"}}},"required":true},"responses":{"200":{"description":"Album added","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"400":{"description":"Please enter a valid name and description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/auth/users":{"get":{"tags":["Auth Controller"],"summary":"List users Api","operationId":"Users","responses":{"403":{"description":"Token/Scope Error","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"201":{"description":"List of Users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"401":{"description":"Please check Access Token","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/auth/profile":{"get":{"tags":["Auth Controller"],"summary":"View Profile","operationId":"profile","responses":{"200":{"description":"View Profile","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}},"403":{"description":"Token/Scope Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}},"401":{"description":"Please check Access Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/albums":{"get":{"tags":["Album Controller"],"summary":"List album Api","operationId":"albums","responses":{"403":{"description":"Token/Scope Error","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"200":{"description":"List of Albums","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"401":{"description":"Please check Access Token","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/albums/{album_id}":{"get":{"tags":["Album Controller"],"summary":"List album by album id","operationId":"albums_by_id","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"403":{"description":"Token/Scope Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"200":{"description":"List of Albums","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"401":{"description":"Please check Access Token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/download-thumbnail":{"get":{"tags":["Album Controller"],"operationId":"downloadThumbnail","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/download-photo":{"get":{"tags":["Album Controller"],"operationId":"downloadPhoto","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/":{"get":{"tags":["home-controller"],"operationId":"demo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/profile/delete":{"delete":{"tags":["Auth Controller"],"summary":"Delete Profile","operationId":"delete_profile","responses":{"403":{"description":"Token/Scope Error","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Please check Access Token","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Delete Password","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/delete":{"delete":{"tags":["Album Controller"],"summary":"delete a  photo","operationId":"delete_photo","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"Photo deleted","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"Spring-Demo-Api":[]}]}},"/api/v1/albums/{album_id}/delete":{"delete":{"tags":["Album Controller"],"summary":"Delete an Album","operationId":"delete_album","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"Album Deleted","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"Spring-Demo-Api":[]}]}}},"components":{"schemas":{"AuthoritiesDTO":{"required":["authorities"],"type":"object","properties":{"authorities":{"type":"string","description":"Authorities","example":"USER"}}},"AccountViewDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"authorities":{"type":"string"}}},"PasswordDTO":{"required":["password"],"type":"object","properties":{"password":{"maxLength":20,"minLength":6,"type":"string","description":"Password","example":"password!@#$34"}}},"AlbumPayloadDTO":{"required":["description","title"],"type":"object","properties":{"title":{"type":"string","description":"Album name","example":"Travel"},"description":{"type":"string","description":"Description of the album","example":"Description"}}},"AlbumViewDTO":{"required":["description","title"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string","description":"Album name","example":"Travel"},"description":{"type":"string","description":"Description of the album","example":"Description"},"photos":{"type":"array","items":{"$ref":"#/components/schemas/PhotoDTO"}}}},"PhotoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"fileName":{"type":"string"},"download_link":{"type":"string"}}},"PhotoPayloadDTO":{"required":["description","title"],"type":"object","properties":{"title":{"type":"string","description":"Photo name","example":"Seflie"},"description":{"type":"string","description":"Description of the Photo","example":"Description"}}},"PhotoViewDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"}}},"AccountDTO":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"Email Address","example":"example@example.com"},"password":{"maxLength":20,"minLength":6,"type":"string","description":"Password","example":"password!@#$34"}}},"UserLoginDTO":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"Email Address","example":"user01@user.com"},"password":{"maxLength":20,"minLength":6,"type":"string","description":"Password","example":"password$01"}}},"TokenDTO":{"type":"object","properties":{"token":{"type":"string"}}},"ProfileDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"authorities":{"type":"string"}}}},"securitySchemes":{"Spring-Demo-Api":{"type":"http","in":"header","scheme":"bearer"}}}}